requirements_frozen.txt
addict==2.4.0
attributee==0.1.9
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.1.1
cycler==0.12.1
dotty-dict==1.3.1
fonttools==4.57.0
idna==3.11
importlib_resources==6.4.5
kiwisolver==1.4.7
lapx==0.9.4
matplotlib==3.7.5
mmcls==0.25.0
mmcv-full==1.5.3
mmdet==2.24.1
mmtrack==0.14.0
motmetrics==1.4.0
numpy==1.24.4
opencv-python==4.13.0.90
packaging==26.0
pandas==1.3.5
pillow==10.4.0
platformdirs==4.3.6
pycocotools==2.0.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.4
scipy==1.7.3
seaborn==0.13.2
six==1.17.0
terminaltables==3.1.10
tomli==2.4.0
torch==1.10.1+cu113
torchvision==0.11.2+cu113
tqdm==4.67.1
urllib3==2.2.3
xmltodict==0.15.0
yapf==0.43.0
zipp==3.20.2
